humphi_search/requirements.txt
# Humphi AI File Search — Requirements
# Install with: pip install -r requirements.txt

# Core vector search
chromadb>=0.4.0
sentence-transformers>=2.2.0

# File watching
watchdog>=3.0.0

# Document extraction
pypdf>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
python-pptx>=0.6.21

# Image metadata
Pillow>=10.0.0

# Cloud backup (optional — only needed if using Supabase)
# supabase>=2.0.0
